Add new file with validator function
