+++ index.html
@@ -8,6 +8,7 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <!--<script src="/js/functions.js"></script>-->
   <script src="/vendor/pristine/pristine.min.js"></script>
+  <script src="/vendor/nouislider/nouislider.js"></script>
   <script src="/js/main.js" type="module"></script>
   <title>Кекстаграм</title>
 </head>
